// client/redux/Chat/chat.action-types.ts
export enum ActionTypes {
  CREATE_CHAT = 'CREATE_CHAT',
  CREATE_CHAT_SUCCESS = 'CREATE_CHAT_SUCCESS',
  CREATE_CHAT_ERROR = 'CREATE_CHAT_ERROR',

  SAVE_CHAT = 'SAVE_CHAT',
  SAVE_CHAT_SUCCESS = 'SAVE_CHAT_SUCCESS',
  SAVE_CHAT_ERROR = 'SAVE_CHAT_ERROR',

  FETCH_CHAT = 'FETCH_CHAT',
  FETCH_CHAT_SUCCESS = 'FETCH_CHAT_SUCCESS',
  FETCH_CHAT_ERROR = 'FETCH_CHAT_ERROR',

}